Conflicts:
	src/auto/configure

+++ src/misc1.c
@@ -1654,7 +1654,7 @@ get_leader_len(line, flags, backward, include_space)
 	    if (vim_iswhite(string[0]))
 	    {
 		if (i == 0 || !vim_iswhite(line[i - 1]))
-		    continue;  /* missing shite space */
+		    continue;  /* missing white space */
 		while (vim_iswhite(string[0]))
 		    ++string;
 	    }
@@ -10145,7 +10145,7 @@ expand_path_option(curdir, gap)
 # if defined(MSWIN) || defined(MSDOS)
 	/* Avoid the path ending in a backslash, it fails when a comma is
 	 * appended. */
-	len = STRLEN(buf);
+	len = (int)STRLEN(buf);
 	if (buf[len - 1] == '\\')
 	    buf[len - 1] = '/';
 # endif
